requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Limiter
Flask-CORS
cloudinary
requests
python-dotenv